requirements.txt
fuzzywuzzy==0.18.0
Pillow==9.5.0
Requests==2.31.0
zhconv==1.4.2
